% Generated by roxygen2 (4.0.2): do not edit by hand
\name{installAllPakcages}
\alias{installAllPakcages}
\title{A routine to install all CRAN packages}
\usage{
installAllPakcages()
}
\arguments{
\item{none}{}
}
\description{
A routine to install all CRAN packages
}
\examples{
installAllPakcages()
}
\keyword{packages}

