% Generated by roxygen2 (4.0.2): do not edit by hand
\name{installPakcages}
\alias{installPakcages}
\title{A routine to install essential packages}
\usage{
installPakcages()
}
\arguments{
\item{none}{}
}
\description{
A routine to install essential packages
}
\examples{
installPakcages()
}
\keyword{packages}

