% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isnull}
\alias{isnull}
\title{Like SQL isnull(), replace first value with second value if first one is NA}
\usage{
isnull(x, y)
}
\arguments{
\item{x}{first value (may be a vector)}

\item{y}{second value (may be a vector)}
}
\description{
Like SQL isnull(), replace first value with second value if first one is NA
}
\examples{
isnull(c(1,NA,1), c(0,0,0))
}
\keyword{null}
\keyword{replace}

